Add support for YouTube URL's with a specific time
Adds an extra variable to catch the 'start_time' data after the 'youtube_dl' module extracts the info. It returns a float of how many seconds in the video should start from which is converted into an int for simplicities sake. If there was a 'start_time' it will append the MPV option "--start=+" plus whatever the number of seconds is. I can only test this on my computer so some testing may be necessary. Also Python's not my main language so I apologize if it's messy or wrong, if so feel free to give me tips or just rewrite it.